Create file to ingest power plant data
When the user runs this file, it downloads and extracts data from a zip file which contains a CSV file of power plants and data about said power plants. The user can specify a url for the file in the arguments or use the default data. The code then converts latitude longitude pairs into a longitude latitude array for use with elasticsearch, and finally creates a new index named powerplants to store all the data.